Add Session Provider to be able to generate manual imap/smtp
